// src/types/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type PublicEnumProofingStatus =
  | "Out for Proof"
  | "Response from Client"
  | "Approved"
  | "Disapproved"
  | "Canceled"
export type PublicEnumTaskPriority = "Low" | "Medium" | "High"
export type PublicEnumTaskStatus = "To-Do" | "In Progress" | "Done" | "On Hold"
export type PublicEnumTeamRole = "Admin" | "Member" | "Viewer"

export interface Database {
  public: {
    Tables: {
// ...
      user_profiles: {
        Row: {
          default_custom_domain: string | null
          elevenlabs_api_key: string | null
          elevenlabs_webhook_url: string | null
          google_api_key: string | null
          google_build_config: Json | null
          id: string
          sendgrid_api_key: string | null
          sendgrid_from_email: string | null
          twilio_account_sid: string | null
          twilio_auth_token: string | null
          twilio_from_number: string | null
          updated_at: string | null
        }
        Insert: {
          default_custom_domain?: string | null
          elevenlabs_api_key?: string | null
          elevenlabs_webhook_url?: string | null
          google_api_key?: string | null
          google_build_config?: Json | null
          id: string
          sendgrid_api_key?: string | null
          sendgrid_from_email?: string | null
          twilio_account_sid?: string | null
          twilio_auth_token?: string | null
          twilio_from_number?: string | null
          updated_at?: string | null
        }
        Update: {
          default_custom_domain?: string | null
          elevenlabs_api_key?: string | null
          elevenlabs_webhook_url?: string | null
          google_api_key?: string | null
          google_build_config?: Json | null
          id?: string
          sendgrid_api_key?: string | null
          sendgrid_from_email?: string | null
          twilio_account_sid?: string | null
          twilio_auth_token?: string | null
          twilio_from_number?: string | null
          updated_at?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      proofing_status: PublicEnumProofingStatus
      task_priority: PublicEnumTaskPriority
      task_status: PublicEnumTaskStatus
      team_role: PublicEnumTeamRole
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

export type Tables<T extends keyof Database["public"]["Tables"]> =
  Database["public"]["Tables"][T]["Row"]
export type TablesInsert<T extends keyof Database["public"]["Tables"]> =
  Database["public"]["Tables"][T]["Insert"]